scraper/requirements.txt
scrapy>=2.8.0
requests>=2.28.0
beautifulsoup4>=4.11.1
python-dotenv>=0.20.0
spacy>=3.4.0
newspaper3k>=0.2.8
chromadb>=0.4.6
wikipedia-api>=0.5.7
newsapi-python>=0.2.6 